Added get_dealer_reviews_from_cf to restapis.py.
Function calls the cloud functions url with a dealer_id to obtain reviews for the dealership. It then breaks each review into individual objects.
It runs these new objects through a function that will analyze the sentiment of the reviews. This review function is not yet implemented and is the next step of development.